+++ src/main/application-constants.ts
@@ -23,7 +23,8 @@ export function windowDefaults() {
     width: screenWidth,
   } = screen.getPrimaryDisplay().workAreaSize
   const headerHeight = 38
-  const defaultHeight = Math.min(700 + headerHeight, screenHeight)
+  // NOTE(maxph): we have a padding of 66px on the left of the buttons
+  const defaultHeight = Math.min(802 + headerHeight, screenHeight)
 
   const x = (screenWidth - defaultWidth) / 2
   const y = (screenHeight - defaultHeight) / 2