+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 ### Changed
-- forward message dialog now has a title and a confirmation dialog before forwarding
+- forward message dialog now has a title and a confirmation dialog and chat preview before forwarding
 
 ### Fixed