+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - forward message dialog now has a title and a confirmation dialog and chat preview before forwarding
+- update deltachat-node and deltachat/jsonrpc-client to `v1.105.0`
 
 ### Fixed
 - fix default welcome screen height so all buttons fit scrollbar is hidden